User-Info App
User-Info App made with TypeScript as a part of Integrify frontend course. This app is fetching the data about GitHub users and provides the requested info to the client. As a result of work on this tasks, the principles of working with fetching data and working with it in TypeScript were studied and worked out.